Create SublistSearch.c
In the sublist search implementation, linked lists are created first using struct keyword in the C language. These linked lists are checked whether they are not empty; and then the elements are compared one by one linearly to find a match. If the second linked list is present in the first linked list in the same order, then the output is returned TRUE; otherwise the output is printed FALSE.